moved IFrameLoadInfo and IFrameSandboxState to script_traits
