+++ components/msg/constellation_msg.rs
@@ -36,12 +36,6 @@ impl<T: Serialize + Deserialize> Clone for ConstellationChan<T> {
     }
 }
 
-#[derive(PartialEq, Eq, Copy, Clone, Debug, Deserialize, Serialize)]
-pub enum IFrameSandboxState {
-    IFrameSandboxed,
-    IFrameUnsandboxed
-}
-
 // We pass this info to various threads, so it lives in a separate, cloneable struct.
 #[derive(Clone, Copy, Deserialize, Serialize)]
 pub struct Failure {
@@ -204,23 +198,6 @@ bitflags! {
         const ALT = 0x04,
         const SUPER = 0x08,
     }
-}
-
-/// Specifies the information required to load a URL in an iframe.
-#[derive(Deserialize, Serialize)]
-pub struct IframeLoadInfo {
-    /// Url to load
-    pub url: Option<Url>,
-    /// Pipeline ID of the parent of this iframe
-    pub containing_pipeline_id: PipelineId,
-    /// The new subpage ID for this load
-    pub new_subpage_id: SubpageId,
-    /// The old subpage ID for this iframe, if a page was previously loaded.
-    pub old_subpage_id: Option<SubpageId>,
-    /// The new pipeline ID that the iframe has generated.
-    pub new_pipeline_id: PipelineId,
-    /// Sandbox type of this iframe
-    pub sandbox: IFrameSandboxState,
 }
 
 #[derive(Deserialize, HeapSizeOf, Serialize)]
