+++ components/script_traits/script_msg.rs
@@ -2,11 +2,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+use IFrameLoadInfo;
 use canvas_traits::CanvasMsg;
 use euclid::point::Point2D;
 use euclid::size::Size2D;
 use ipc_channel::ipc::IpcSender;
-use msg::constellation_msg::{AnimationState, DocumentState, IframeLoadInfo, NavigationDirection};
+use msg::constellation_msg::{AnimationState, DocumentState, NavigationDirection};
 use msg::constellation_msg::{Failure, MozBrowserEvent, PipelineId};
 use msg::constellation_msg::{LoadData, SubpageId};
 use msg::constellation_msg::{MouseButton, MouseEventType};
@@ -72,7 +73,7 @@ pub enum ScriptMsg {
     /// Notification that this iframe should be removed.
     RemoveIFrame(PipelineId),
     /// A load has been requested in an IFrame.
-    ScriptLoadedURLInIFrame(IframeLoadInfo),
+    ScriptLoadedURLInIFrame(IFrameLoadInfo),
     /// Requests that the constellation set the contents of the clipboard
     SetClipboardContents(String),
     /// Mark a new document as active
